Move Github urls to separate module to introduce method
