+++ src/databricks/labs/ucx/github.py
@@ -1,0 +1,6 @@
+DOCS_URL = "https://databrickslabs.github.io/ucx/docs/"
+GITHUB_URL = "https://github.com/databrickslabs/ucx"
+
+
+def construct_new_issue_url() -> str:
+    return f"{GITHUB_URL}/issues/new"
